Vs in company

+++ LeetCode/BinaryTreeLevelOrderTraversal.cpp
@@ -17,7 +17,7 @@ public:
 		if (!node)
 			return;
 
-		if (depth >= out.size())
+		if (depth >= (int)out.size())
 			out.push_back(vector<int>());
 
 		out[depth].push_back(node->val);
